+++ src/lang/cpp.ts
@@ -1,3 +1,4 @@
+import { AiXCompletionItem } from "../extension";
 import { ID_REGEX, LangUtil } from "./langUtil";
 
 const keywords = new Set<string>();
@@ -63,5 +64,12 @@ export class CppLangUtil extends LangUtil {
         } else {
             return false;
         }
+    }
+
+    public retrigger(completionItem: AiXCompletionItem) {
+        if (completionItem.label.endsWith("->") || completionItem.label.endsWith("::")) {
+            return true;
+        }
+        return super.retrigger(completionItem);
     }
 }